Separted Citations, added terminal dectection
Separted the resources used from the README.md to RESOURCES.md. Starting working on creating a dynamically allocated array from the values of the terminal screen. Using an extern crate for that and using colorterm to detect the key evets from the keyboard. Ignoring all input from the Mouse as it's going to be a keyboard based game.